change console structure

+++ app/Console/Commands/RetireOffer.php
@@ -4,6 +4,7 @@ namespace App\Console\Commands;
 
 use App\Models\Offer;
 use App\Services\SlackService;
+use App\WorkerClasses\HelperFunctions;
 use Illuminate\Console\Command;
 
 class RetireOffer extends Command
@@ -27,22 +28,24 @@ class RetireOffer extends Command
      */
     public function handle()
     {
-        // retire all offers that have passed their expiration date and their robosatsId is less than 20000
-        $offers = Offer::where([['expires_at', '<', now()], ['robosatsId', '<', 100000]])->orWhere([['status', '=', 14], ['robosatsId', '<', 50000]])->get();
-        foreach ($offers as $offer) {
+        if ((new HelperFunctions())->slackCommandCheck()) {
+            // retire all offers that have passed their expiration date and their robosatsId is less than 20000
+            $offers = Offer::where([['expires_at', '<', now()], ['robosatsId', '<', 100000]])->orWhere([['status', '=', 14], ['robosatsId', '<', 50000]])->get();
+            foreach ($offers as $offer) {
 
-            $randomNumber = rand(5000000, 10000000);
-            $offer->robosatsIdStorage = $offer->robosatsId;
-            $offer->robosatsId = $randomNumber;
-            $offer->save();
+                $randomNumber = rand(5000000, 10000000);
+                $offer->robosatsIdStorage = $offer->robosatsId;
+                $offer->robosatsId = $randomNumber;
+                $offer->save();
 
-            // if there is a slack channel associated with the offer, archive it
-            $slackService = new SlackService();
-            $slackService->deleteChannel($offer->slack_channel_id);
+                // if there is a slack channel associated with the offer, archive it
+                $slackService = new SlackService();
+                $slackService->deleteChannel($offer->slack_channel_id);
 
-            $offer->slack_channel_id = null;
-            $offer->save();
+                $offer->slack_channel_id = null;
+                $offer->save();
 
+            }
         }
     }
 }
